+++ src/Simplify.Repository.EntityFramework/UnitOfWork.cs
@@ -6,15 +6,14 @@ namespace Simplify.Repository.EntityFramework;
 /// <summary>
 /// Provides unit of work
 /// </summary>
-public class UnitOfWork<T> : IUnitOfWork
+/// <remarks>
+/// Initializes a new instance of the <see cref="UnitOfWork{T}" /> class.
+/// </remarks>
+/// <param name="context">The context.</param>
+/// <exception cref="InvalidOperationException">Error opening session, session is null</exception>
+public class UnitOfWork<T>(DbContext context) : IUnitOfWork
 	where T : DbContext
 {
-	/// <summary>
-	/// Initializes a new instance of the <see cref="UnitOfWork{T}" /> class.
-	/// </summary>
-	/// <param name="context">The context.</param>
-	/// <exception cref="System.InvalidOperationException">Error opening session, session is null</exception>
-	public UnitOfWork(DbContext context) => Context = context ?? throw new InvalidOperationException("Error opening session, session is null");
 
 	/// <summary>
 	/// Gets the session.
@@ -22,7 +21,7 @@ public class UnitOfWork<T> : IUnitOfWork
 	/// <value>
 	/// The session.
 	/// </value>
-	public DbContext Context { get; }
+	public DbContext Context { get; protected set; } = context ?? throw new InvalidOperationException("Error opening session, session is null");
 
 	/// <summary>
 	/// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.