+++ src/Simplify.Repository.EntityFramework/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Dependencies
 
+- Simplify.EntityFramework bump to 1.0
 - Simplify.Repository bump to 1.7
 - Microsoft.EntityFrameworkCore.Relational bump to 5.0.17
 - Simplify.DI bump to 4.2.10